requirements.txt
discord.py>=2.4.0
asyncpg>=0.30.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
